requirements.txt
requests==2.28.1
gitpython==3.1.31
